+++ gendarme/rules/AltCode.Rules.PowerShell/UseOnlyStandardVerbsRule.fs
@@ -1,6 +1,7 @@
 ﻿namespace AltCode.Rules.PowerShell
 
 open System
+open System.Globalization
 
 open Mono.Cecil
 open Mono.Cecil.Cil
@@ -13,9 +14,9 @@ open System.Diagnostics.CodeAnalysis
 
 [<Problem("The namespace in which the cmdlet class is defined must have a name in the following format: '<Product>.Commands'.>")>]
 [<Solution("Move or rename the class to a namespace that identifies the product and ends in '.Commands'.")>]
-[<FxCopCompatibility("Microsoft.PowerShell", "PS1011:DefineCmdletInTheCorrectNamespace")>]
+[<FxCopCompatibility("Microsoft.PowerShell", "PS1001:UseOnlyStandardVerbs")>]
 [<Sealed>]
-type DefineCmdletInTheCorrectNamespaceRule() =
+type UseOnlyStandardVerbsRule() =
   inherit Rule()
 
   interface ITypeRule with
@@ -26,10 +27,19 @@ type DefineCmdletInTheCorrectNamespaceRule() =
       let td = ``type``
 
       if Tools.IsCmdlet td then
-        if td.Namespace.EndsWith(".Commands", StringComparison.Ordinal) |> not
+        let attr = td.CustomAttributes
+                   |> Seq.find (fun a -> a.AttributeType.Inherits Tools.cmdletAttribute)
+        let verb = (attr.ConstructorArguments |> Seq.head).Value.ToString()
+        if Tools.standardVerbs
+           |> Seq.exists (fun v -> v.Equals(verb, StringComparison.OrdinalIgnoreCase))
+           |> not
         then
-          this.Runner.Report(td, Severity.High, Confidence.High, td.FullName)
+          let msg = String.Format(CultureInfo.InvariantCulture,
+                                  "Non-standard verb {0} used here.",
+                                  verb)
+          this.Runner.Report(td, Severity.High, Confidence.High, msg)
 
         this.Runner.CurrentRuleResult
+
       else
         RuleResult.DoesNotApply