+++ ReleaseNotes.md
@@ -1,6 +1,7 @@
 # 2022.x.x.xxxxx-pre-release 
 
 * New rules
-  * `AltCode.Rules.General.PreferStrongNamedAssembliesRule` -- to replace deprecated/withdrawn FxCop rule Microsoft.Design#CA2210
+  * `AltCode.Rules.General.PreferStrongNamedAssembliesRule` to replace deprecated/withdrawn FxCop rule Microsoft.Design#CA2210
+  * `AltCode.Rules.PowerShell.UseOnlyStandardVerbsRule` to replace "Microsoft.PowerShell#PS1001:UseOnlyStandardVerbs"
 
 ⁋For previous releases, go here -- https://github.com/SteveGilham/Gendarme/blob/trunk/ReleaseNotes%20-%20Previously.md